Fix CronJob schedule for in webhook chart
The current schedule (* * */24 * *) effectively causes the CronJob to
run on the 24th of every month only, and on the 24th it runs every
single minute (1440 times). The itention was to have it run daily, but I
suspect that running it weekly should be plenty good enough.

This fixes helm/charts#8077.

Signed-off-by: Chris Boot <bootc@bootc.net>