[elasticsearch] Tweak the 'readinessProbe' command to verify that master nodes are available.
The behaviour of the `/` endpoint changed[0] between 6.x and 7.x, whereby
previously it would return a HTTP `503` response when the cluster was
blocked, it now returns a HTTP `200` response even if there are no masters
available.

This change updates the behaviour of the `readinessProbe` command during
normal running to verify that the local node is responding and that there
are master nodes available. [1]

The desired behaviour here is that if the data nodes are unable to talk to
their master nodes for whatever reason, then the data nodes will become
`Unready` and therefore be removed from the Service load-balancer until
the master nodes are available again.

Refs:
[0] https://github.com/elastic/elasticsearch/pull/29045
[1] https://github.com/elastic/elasticsearch/issues/34897#issuecomment-519441101