Fix problems with Thanos Ruler statefulset
This commit fixes problems that occur when deploying Thanos Ruler using the BanzaiCloud Thanos Helm chart.

Problems identified when deploying on Helm 3 are:
- misuse of the field "resources" in StatefulSet.spec.template.spec
- misuse of the field "secretName" in StatefulSet.spec.template.spec
- omission of the field "serviceName" in StatefulSet.spec

The present commit moves "resources" up to the StatefulSet.spec.template.spec.containers field, where it is valid. It removes "secretName" entirely, since it does not exist anywhere in the StatefulSet API. Finally, it adds a serviceName, using the "thanos.name" value set.

There should be no problems deploying this same commit on any Helm under version 3, since there is no evidence that any Kubernetes version prior to the most recent version has these fields in the places where they can be found in this template. It is believed that previous versions of Helm were simply not reporting the error and allowing deployment to continue.