[elasticsearch] use bash for keystore init container (#1464)
This commit is a follow up of fb9f444d and is updating the keystore
initContainer to also use bash instead of sh. This is required for
Elasticsearch > 7.16.0 because the Docker image is now based on Ubuntu
instead of CentOS 8, and sh on Ubuntu isn't compatible with the
`if [[... -eq ... ]]` statements.